[task-13] useExcelFile 생성 및 MissingData 수정
