+++ src/components/MissingData/MissingData.module.scss
@@ -1,3 +1,12 @@
 .wrapper {
   padding: 10px;
+
+  .message + .message {
+    margin-top: 10px;
+  }
+
+  .message {
+    color: red;
+    font-weight: 600;
+  }
 }
