+++ src/hooks/useExcelFile/index.ts
@@ -1,0 +1,4 @@
+import { FileChangeEvent } from './useExcelFile';
+
+export { default } from './useExcelFile';
+export type { FileChangeEvent };
